requirements.txt
fastapi~=0.73.0
neomodel~=4.0.8
python-dotenv~=0.19.2
pydantic~=1.9.0
requests~=2.27.1
web3~=5.28.0
polling2~=0.5.0
APScheduler~=3.8.1
uvicorn~=0.17.4
locust~=2.7.3